+++ processor/transform/readme-to-mdx.ts
@@ -22,7 +22,7 @@ const readmeToMdx = (): Transform => tree => {
 
   visit(tree, 'rdme-callout', (node, index, parent) => {
     const isEmpty = node.data.hProperties?.empty;
-    const isH3 = node.children[0].type === 'heading' && node.children[0].depth === 3;
+    const isH3 = node.children?.[0] && 'type' in node.children[0] && node.children[0].type === 'heading' && node.children[0].depth === 3;
     let { icon, theme } = node.data.hProperties;
     if (!icon) icon = defaultIcons[theme];
     if (!theme) theme = themes[icon] || 'default';
@@ -77,7 +77,6 @@ const readmeToMdx = (): Transform => tree => {
 
   visit(tree, 'figure', (figure, index, parent) => {
     const [image, caption] = figure.children;
-
     const { align, width } = image.data.hProperties;
     const border = image.data.hProperties.className === 'border';