+++ processor/transform/tables-to-jsx.ts
@@ -34,6 +34,9 @@ const visitor = (table: Table, index: number, parent: Parents) => {
         ? (cell.children[0] as unknown as Paragraph).children[0]
         : cell.children[0];
 
+    // If cell is empty there's nothing to visit
+    if (!content) return EXIT;
+
     // @note: Compatibility with RDMD. Ideally, I'd put this in a separate
     // transformer, but then there'd be some duplication.
     visit(cell, 'break', (_, breakIndex, breakParent) => {